backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "bigchaipat", email = "bchaipats@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi[standard]>=0.128.0",
    "greenlet>=3.3.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
    "sqlmodel>=0.0.31",
]

[project.scripts]
backend = "backend:main"

[build-system]
requires = ["uv_build>=0.9.11,<0.10.0"]
build-backend = "uv_build"
